Implementando CRUD para payment-methods, product e category

+++ src/services/category.service.ts
@@ -1,13 +1,17 @@
 import { NotFoundError } from "../errors/not-found.error.js";
+import { ValidationErro } from "../errors/validation.error.js";
 import { Category } from "../models/category.model.js";
 import { CategoryRepository } from "../repositories/category.repository.js";
+import { ProductRepository } from "../repositories/product.repository.js";
 
 export class CategoryService {
 
     private _repository: CategoryRepository;
+    private _productRepository: ProductRepository;
 
     constructor() {
         this._repository = new CategoryRepository();
+        this._productRepository = new ProductRepository();
     }
 
     async getAll(): Promise<Category[]> {
@@ -37,6 +41,9 @@ export class CategoryService {
     }
 
     async delete(id: string) {
+        if (await this._productRepository.getCountByCategory(id) > 0)
+            throw new ValidationErro("Não é possível excluir uma categoria com produtos relacionados");
+
         await this._repository.delete(id);
     }
 }
